requirements.txt
# HYDRA Competitive Intelligence System
# Core dependencies

# FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10

# HTTP and web scraping
httpx==0.28.1
beautifulsoup4==4.13.5

# Data processing
pandas==2.3.2
numpy==2.3.2

# AI and ML
ollama==0.5.3
chromadb==1.0.20

# Utilities
python-dotenv==1.1.1
pydantic==2.11.7

# Development and testing (optional)
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.3.0
flake8==6.0.0
